# Undo grid changes to reshape_and_cache