--- a/tools/benchmarks/collect_reshape_and_cache.sh
+++ b/tools/benchmarks/collect_reshape_and_cache.sh
@@ -2,10 +2,10 @@
 # Copyright 2025 Stack AV Co.
 # SPDX-License-Identifier: Apache-2.0
 
+# Specify CONCH_ENABLE_VLLM=1 to compare against vLLM ref impl (if installed)
 # Specify CONCH_BENCH_NO_CSV=1 to print results to stdout instead of file
 
 # Need to enable vLLM to compare against vLLM CUDA implementation
-export CONCH_ENABLE_VLLM=1
 export VLLM_LOGGING_LEVEL=CRITICAL
 
 # Create output directory
@@ -14,11 +14,12 @@ benchmark_dir="results/$benchmark_name"
 mkdir -p $benchmark_dir
 
 num_tokens=(
-  "32"
   "64"
   "128"
-  "256"
   "512"
+  "2048"
+  "8192"
+  "32768"
 )
 
 for tokens in ${num_tokens[@]}; do